requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-markdownx==4.0.7
django-slugger==1.1.2
Markdown==3.6
pillow==10.4.0
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
Unidecode==1.3.8
